Lexify straight in to pratt.Token
