+++ frontend/src/metabase-lib/v1/expressions/pratt/fuzz.unit.spec.ts
@@ -1,16 +1,15 @@
 import process from "process";
 import _ from "underscore";
 
+import { lexify } from "../pratt/lexifier";
 import { generateExpression } from "../test/generator";
-
-import { tokenize } from "./tokenize";
 
 const fuzz = process.env.MB_FUZZ ? describe : _.noop;
 
 describe("metabase-lib/v1/expressions/tokenizer", () => {
   // quick sanity check before the real fuzzing
   it("should tokenize custom expresssion", () => {
-    expect(() => tokenize("CASE([Deal],[Price]*7e-1,[Price]")).not.toThrow();
+    expect(() => lexify("CASE([Deal],[Price]*7e-1,[Price]")).not.toThrow();
   });
 });
 
@@ -20,12 +19,12 @@ fuzz("FUZZING metabase-lib/v1/expressions/tokenizer", () => {
   for (let seed = 0; seed < MAX_SEED; ++seed) {
     it("should handle generated expression from seed " + seed, () => {
       const { expression } = generateExpression(seed);
-      expect(() => tokenize(expression)).not.toThrow();
+      expect(() => lexify(expression)).not.toThrow();
     });
 
     it("should not error on generated expression from seed " + seed, () => {
       const { expression } = generateExpression(seed);
-      expect(tokenize(expression).errors).toEqual([]);
+      expect(lexify(expression).errors).toEqual([]);
     });
   }
 });
