+++ frontend/src/metabase-lib/v1/expressions/pratt/syntax.ts
@@ -262,3 +262,18 @@ for (const [key, value] of Object.entries(ALL_NODES)) {
 
 SUB.name = "SUBTRACT";
 WS.name = "WHITESPACE";
+
+export const OPERATORS = new Set([
+  COMMA,
+  GROUP,
+  GROUP_CLOSE,
+  ADD,
+  SUB,
+  MULDIV_OP,
+  EQUALITY,
+  COMPARISON,
+  LOGICAL_AND,
+  LOGICAL_OR,
+  LOGICAL_NOT,
+  NEGATIVE,
+]);
