+++ frontend/src/metabase-lib/v1/expressions/tokenizer/index.ts
@@ -1,2 +1,2 @@
 export { tokenize } from "./tokenize";
-export { OPERATOR, TOKEN } from "./types";
+export { OPERATOR } from "./types";
